@@ -1,5 +1,5 @@
 {
-  "name": "food",
+  "name": "with-apollo-app",
   "author": "Yan Takushevich",
   "license": "ISC",
   "version": "1.0.0",
@@ -13,18 +13,22 @@
   "dependencies": {
     "apollo-boost": "^0.1.3",
     "apollo-link-context": "^1.0.8",
+    "apollo-link-error": "^1.1.1",
     "apollo-link-http": "^1.5.4",
+    "apollo-link-rest": "^0.4.4",
+    "change-case": "^3.0.2",
     "cookie": "^0.3.1",
     "date-fns": "^1.29.0",
+    "fetch-headers": "^2.0.0",
     "graphql": "0.13.2",
     "immer": "^1.5.0",
     "isomorphic-unfetch": "2.0.0",
     "lodash": "^4.17.10",
-    "next": "^6.1.1",
+    "next": "^7.0.1",
     "next-routes": "^1.4.2",
     "node-mailjet": "^3.2.1",
     "react": "^16.4.1",
-    "react-apollo": "2.1.0",
+    "react-apollo": "2.1.11",
     "react-dom": "^16.4.1",
     "react-visibility-sensor": "^3.11.1",
     "semantic-ui-css": "^2.3.3",
@@ -32,23 +36,22 @@
     "validator": "^10.5.0"
   },
   "devDependencies": {
-    "@types/next": "^6.1.0",
+    "@divyenduz/ts-graphql-plugin": "^0.1.0",
+    "@types/next": "^7.0.1",
     "@types/react": "^16.4.7",
-    "@types/styled-jsx": "^2.2.3",
-    "@zeit/next-css": "^0.2.0",
-    "@zeit/next-typescript": "^1.1.0",
-    "autoprefixer": "^9.1.1",
-    "file-loader": "^1.1.11",
+    "@zeit/next-css": "^1.0.1",
+    "@zeit/next-typescript": "^1.1.1",
+    "autoprefixer": "^9.1.5",
+    "file-loader": "^2.0.0",
     "lint-staged": "^7.2.0",
     "prettier": "^1.14.0",
-    "styled-jsx-plugin-postcss": "^0.1.3",
     "tailwindcss": "^0.6.4",
     "ts-node": "^7.0.1",
     "tslint": "^5.11.0",
     "tslint-config-standard": "^7.1.0",
     "tslint-plugin-prettier": "^1.3.0",
     "typescript": "^3.0.1",
-    "url-loader": "^1.0.1"
+    "url-loader": "^1.1.1"
   },
   "lint-staged": {
     "*.{js, tsx}": [
