@@ -1,37 +1,30 @@
 import React, { Component, Fragment } from 'react'
 import { Card } from 'semantic-ui-react'
 
-import Meetup from './item'
+import Video from './item'
 import Preloader from '../Preloader'
 
 interface Props {
   data?: [
     {
-      id: string
-      title: string
-      organizer: {
-        name: string
-      }
-      location: string
-      date: string
-      attendees: [{}]
+      ytId: string
     }
   ]
   loading: boolean
 }
 
-export default class Meetups extends Component<Props> {
+export default class Videos extends Component<Props> {
   render() {
     const { data, loading } = this.props
     if (loading) return <Preloader />
     if (!data) return null
     return (
       <Fragment>
-        <h2 className="ui header">All Meetups</h2>
+        <h2 className="ui header">All Videos</h2>
         <Card.Group itemsPerRow={2}>
           {data.length
-            ? data.map(meetup => <Meetup key={meetup.id} meetup={meetup} />)
-            : 'No meetups yet'}
+            ? data.map(video => <Video key={video.ytId} video={video} />)
+            : 'No videos yet'}
         </Card.Group>
       </Fragment>
     )
