@@ -1,6 +1,8 @@
-import { ApolloClient, InMemoryCache } from 'apollo-boost'
+import { ApolloClient, ApolloLink, InMemoryCache } from 'apollo-boost'
 import { createHttpLink } from 'apollo-link-http'
 import { setContext } from 'apollo-link-context'
+import { RestLink } from 'apollo-link-rest'
+import { onError } from 'apollo-link-error'
 import fetch from 'isomorphic-unfetch'
 
 let apolloClient = null
@@ -10,10 +12,19 @@ if (!process.browser) {
   global.fetch = fetch
 }
 
-function create (initialState, { getToken }) {
+// Polyfill Headers for RestLink
+global.Headers = global.Headers || require('fetch-headers')
+
+const restLink = new RestLink({
+  uri: `https://www.googleapis.com/youtube/v3/videos?part=snippet&key=${
+    process.env['YOUTUBE_API_KEY']
+  }`,
+})
+
+function create(initialState, { getToken }) {
   const httpLink = createHttpLink({
     uri: 'http://localhost:4000',
-    credentials: 'same-origin'
+    credentials: 'same-origin',
   })
 
   const authLink = setContext((_, { headers }) => {
@@ -21,21 +32,31 @@ function create (initialState, { getToken }) {
     return {
       headers: {
         ...headers,
-        authorization: token ? `Bearer ${token}` : ''
-      }
+        authorization: token ? `Bearer ${token}` : '',
+      },
     }
   })
+
+  const errorLink = onError(({ graphQLErrors }) => {
+    if (graphQLErrors) graphQLErrors.map(({ message }) => console.log(message))
+  })
+
+  const links = [authLink, restLink, httpLink]
+
+  if (process.browser) {
+    links.unshift(errorLink)
+  }
 
   // Check out https://github.com/zeit/next.js/pull/4611 if you want to use the AWSAppSyncClient
   return new ApolloClient({
     connectToDevTools: process.browser,
     ssrMode: !process.browser, // Disables forceFetch on the server (so queries are only run once)
-    link: authLink.concat(httpLink),
-    cache: new InMemoryCache().restore(initialState || {})
+    link: ApolloLink.from(links),
+    cache: new InMemoryCache().restore(initialState || {}),
   })
 }
 
-export default function initApollo (initialState, options) {
+export default function initApollo(initialState, options) {
   // Make sure to create a new client for every server-side request so that data
   // isn't shared between connections (which would be bad)
   if (!process.browser) {
