@@ -1,20 +1,19 @@
 import React, { Component, Fragment } from 'react'
 import { Mutation, Query } from 'react-apollo'
 import { gql } from 'apollo-boost'
-import { format } from 'date-fns'
 import { Button, Icon } from 'semantic-ui-react'
 import get from 'lodash/get'
 
 import Preloader from '../components/Preloader'
 
 interface Props {
-  createMeetup: Function
+  addVideo: Function
   query: {
     id: String
   }
 }
 
-export default class CreateMeetup extends Component<Props> {
+export default class CreateVideo extends Component<Props> {
   static getInitialProps({ query }) {
     return { query }
   }
@@ -28,8 +27,8 @@ export default class CreateMeetup extends Component<Props> {
           if (loading) {
             return <Preloader />
           }
-          const { currentUser, meetup } = data
-          const attending = meetup.attendees.some(item => {
+          const { currentUser, video } = data
+          const attending = video.attendees.some(item => {
             return get(currentUser, ['id']) === item.id
           })
           return (
@@ -37,24 +36,20 @@ export default class CreateMeetup extends Component<Props> {
               mutation={mutation}
               variables={{ id, attending: !attending }}
             >
-              {(attendMeetup: Function, { loading }) => (
+              {(attendVideo: Function, { loading }) => (
                 <div>
                   <h1 className="ui dividing header">
-                    {meetup.title}
+                    {video.title}
                     <div className="sub header">
-                      Organized by {meetup.organizer.name}
+                      Organized by {video.organizer.name}
                     </div>
                   </h1>
                   <div className="description">
                     <h3 className="ui header">Details</h3>
-                    <p>{meetup.description}</p>
-                    <p>
-                      <i className="calendar icon" />{' '}
-                      {format(meetup.date || new Date(), 'DD.MM.YYYY, H:m')}
-                    </p>
+                    <p>{video.description}</p>
                     <p>
                       <i className="map marker alternate icon" />{' '}
-                      {meetup.location}
+                      {video.location}
                     </p>
                   </div>
                   {currentUser && (
@@ -64,7 +59,7 @@ export default class CreateMeetup extends Component<Props> {
                         primary={attending}
                         icon
                         onClick={() =>
-                          attendMeetup({ id, attending: !attending })
+                          attendVideo({ id, attending: !attending })
                         }
                         title={attending ? "I'm not going" : "I'm going"}
                         loading={loading}
@@ -74,11 +69,11 @@ export default class CreateMeetup extends Component<Props> {
                     </Fragment>
                   )}
                   <h3 className="ui header">
-                    Attendees: {meetup.attendees.length}
+                    Attendees: {video.attendees.length}
                   </h3>
                   <div className="ui bulleted list">
-                    {meetup.attendees.length
-                      ? meetup.attendees.map(attendee => (
+                    {video.attendees.length
+                      ? video.attendees.map(attendee => (
                           <div key={attendee.id} className="item">
                             {attendee.name}
                           </div>
@@ -96,21 +91,13 @@ export default class CreateMeetup extends Component<Props> {
 }
 
 const query = gql`
-  query MeetupQuery($id: ID!) {
+  query VideoQuery($ytId: String!) {
     currentUser {
       id
     }
-    meetup(id: $id) {
+    video(ytId: $ytId) {
       id
-      title
-      description
-      date
-      location
-      organizer {
-        name
-      }
-      attendees {
-        id
+      adder {
         name
       }
     }
@@ -118,20 +105,9 @@ const query = gql`
 `
 
 const mutation = gql`
-  mutation AttendMeetupMutation($id: ID!, $attending: Boolean!) {
-    attendMeetup(id: $id, attending: $attending) {
-      id
-      title
-      description
-      date
-      location
-      organizer {
-        name
-      }
-      attendees {
-        id
-        name
-      }
+  mutation BookmarkVideoMutation($ytId: String!, $adding: Boolean!) {
+    bookmarkVideo(ytId: $ytId, adding: $adding) {
+      ytId
     }
   }
 `
